Better fix for the way migrations are loaded.  No symlinks needed anymore.
The spec_helper was explicitly loading the contents of the dummy migrate directory.  So, just add a new line that explicitly loads the contents of the gem's migrate directory as well.  Specs continue to pass, and all migrations will now be automatically loaded without any extra work on the developer's part.