Fix foreign key creation
Migrations don't work with last master because after namespacing
mailboxer (6e7e042cc8c44958688c686d95f84373a42519e7), column to be used
by the foreign key is not correctly guessed anymore. This was not
detected by the test suite because it uses sqlite and foreign keys are
probably ignored by default in sqlite. It's failing in postgresql
though. This commit fixes it by manually expliciting the column.