Remove duplicated migration
This migration was "merged" into main migration.
(20110511145103_create_mailboxer.rb) in commit
6e7e042cc8c44958688c686d95f84373a42519e7, so it now fails with a
duplicate column error. Test suite is not failing because the migration
was correctly removed from the test suite in that commit.

This makes me think that you shouldn't probably maintain two sets of
duplicated files in the repo, because bugs like this one are more likely
to be introduced.